rename auth plugin class.
